feat: test&train data/detector check
